PR/requirements.txt
streamlit==1.28.1
pandas==2.0.3
plotly==5.17.0
mysql-connector-python==8.1.0
requests==2.31.0 